Crate pabst

Point ABSTraction library in Rust.

This library is inspired by PDAL, which provides both point abstraction and a wide range of other abilitiles. Pabst hopes to keep its focus more narrow. We provide source and sink abstractions for LiDAR point cloud file and stream formats. Other libraries should be used to provide intermediate functionality.

Reexports

pub use error::Error;

pub use point::Point;

pub use source::{open_file_source, FileSource, Source};

pub use sink::{open_file_sink, FileSink, Sink};

Modules

error
Our error enum.

point
A catchall LiDAR point.

sink
Point sinks.

source
Sources of points.

Type Definitions

Result
Our custom result type.