Struct pabst::point::Intensity

pub struct Intensity {
    // some fields omitted
}

An intensity value, with information about its minimum and maximum possible values.

Methods

impl Intensity

fn from_u16(value: u16) -> Intensity

Creates a new intensity with the min and the max set to the u16 bounds.

Examples

use pabst::point::Intensity;
let intensity = Intensity::from_u16(10);

fn new(value: f64, min: f64, max: f64) -> Intensity

Creates a new intensity with the value, minimu, and maximum.

Examples

use pabst::point::Intensity;
let intensity = Intensity::new(1.0, 0.0, 2.0);
assert_eq!(32767, intensity.as_u16());

fn as_u16(&self) -> u16

Returns this intensity value as a u16.

Examples

use pabst::point::Intensity;
let intensity = Intensity::from_u16(10);
assert_eq!(10, intensity.as_u16());