Struct pabst::point::Point

pub struct Point {
    pub x: f64,
    pub y: f64,
    pub z: f64,
    pub intensity: Intensity,
    pub return_number: Option<usize>,
    pub number_of_returns: Option<usize>,
    pub scan_direction: ScanDirection,
    pub edge_of_flight_line: bool,
    pub classification: u8,
    pub synthetic: bool,
    pub key_point: bool,
    pub withheld: bool,
    pub scan_angle: Option<f64>,
    pub point_source_id: Option<u16>,
    pub user_data: Option<u8>,
    pub gps_time: Option<f64>,
    pub range: Option<f64>,
    pub width: Option<f64>,
    pub rg_index: Option<f64>,
    pub facet_number: Option<u8>,
    pub target_type: Option<u8>,
    pub high_channel: Option<bool>,
}

A point.

Fields

x

The x dimension of a point. All points must have these.

y

The y dimension of a point. All points must have these.

z

The z dimension of a point. All points must have these.

intensity

This is the intensity of the backscattered energy, as detected by the LiDAR sensor. Since each sensor has its own dynamic range, an intensity value is actually three values — the intensity value itself, the minimum possible value, and the maximum possible value.

return_number

The return number of this point.

number_of_returns

The total number of returns in the pulse of this point.

scan_direction

The direction that the scanner mirror was travelling when the point was collected. The only reason we include this field is that it is in the las standard. Anyone that doesn't collect scan direction information should just keep this at the default of ScanDirection::Unknown, and let the las sink deal with it.

edge_of_flight_line

Is this point at the edge of a flight line? Again, a crappy field that was included just becuase las.

classification

The classification of this point.

I'm not sure if I want to really define out the ASPRS classification mappings in this library, or leave that to the las library. I'm leaning towards keeping that in las just because I don't like classifications as a general concept, and so I try to stay away from them whenever I can.

The value of 0 is "Created, never classified" according to at least one of the las specifications, so this is a sensible default value.

synthetic

Is this a synthetic point? Urgh. las again. Just keep this false unless someone out there is actually using the synthetic field.

key_point

Okay, las, model key point? Really? You're taking up a perfectly good bit in every single las point just for this field? Really?

withheld

It's not even worth complaining anymore. Just look at a las standard to see what all these crappy fields mean. And while you're at it, try clicking on some of the URLs in the las standards. I'll bet some of them are broken. That's how good ASPRS are (is?).

scan_angle

Whew, back into the normal realm of fields. This is the angle of the mirror when the point was captured. Angle is usually defined with zero being straight out of the scanner.

Las, in all of its wisdom, only stores these as integers, but more sensible formats allow for finer resolution, so we dial this up to f64 and let the las sink throw away perfectly good data.

point_source_id

Why, las, why. What is this, even.

user_data

Okay, I'm done complaining now. In all seriousness, I thought about throwing this field out, in particular, but I figure since las is so used I should just mirror it. But that doesn't mean I can't whine about it.

gps_time

GNSS synchronized time.

range

The range of the point from the scanner. Note that for some formats, this explicity does not equal the cartesian distance to the x, y, z location of the point. This is a bit of an esoteric field but we include it because Riegl uses it a decent amount.

width

The width of the received pulse. Again, a pretty esoteric field, this time from the Riegl side of things. I get thinking more and more how to deal with these strange fields like this.

rg_index

Man, this is the range index, and I have no idea what it means.

facet_number

The number of the mirror facet that was used to send this pulse.

target_type

The target type, as an integer. Riegl uses some coded values for target type (Gaussian, Center of Gravity, etc), but at this point we don't choose to propagate those values up to this level.

high_channel

Is this point from a high channel?

TODO this field should be turned into a newtype or enum