Module pabst::source

Sources of points.

These don't necessarily have to be file format readers, but they usually are.

Modules

las
Source las points.

rxp
Get points from an .rxp file or an rxp stream.

sdc
Implement source for .sdc, Riegl's simple file-based discrete return format.

sdf
Read points from .sdf full waveform files.

Traits

FileSource
A point source that can be opened from a path.

Source
A point source.

Functions

open_file_source
Opens a file source with the given options.